Rename 'models' directory to 'types'
