+++ src/common/components/NarrativeList/NarrativeViewItem.test.tsx
@@ -1,6 +1,6 @@
 import NarrativeViewItem from './NarrativeViewItem';
 import { screen, render } from '@testing-library/react';
-import { NarrativeListDoc } from '../../models/NarrativeDoc';
+import { NarrativeListDoc } from '../../types/NarrativeDoc';
 
 const testDoc: NarrativeListDoc = {
   timestamp: 8032329432,
