+++ src/common/types/NarrativeDoc.tsx
